cleanup old code. Added energy bin width f'n. Made clearer variable names. Hard coded to removed chimney/support ring in OHCal.

+++ calibrations/calorimeter/calo_tower_slope/LiteCaloEval.h
@@ -118,7 +118,7 @@ class LiteCaloEval : public SubsysReco
   float fitmin{0.};
   float fitmax{0.};
 
-  float binwidth{0.001};
+  double binwidth{0.001};
 
   std::string _caloname;
   std::string _filename;
